# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Birthmark API",
    "version": "1.0.0",
    "description": "API documentation using swagger for Birthmark API"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a user with email",
        "description": "Sends a verification link to the provided email address.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com",
                    "description": "The email address to register."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification link will be sent to the given email."
          },
          "400": {
            "description": "Email is invalid or already registered."
          }
        }
      }
    },
    "/api/auth/verify": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Verify user via token",
        "description": "Verify user via token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "token"
                ],
                "properties": {
                  "token": {
                    "type": "string",
                    "format": "string",
                    "example": "aaaaabbbbbbccccc",
                    "description": "Token for verifying email as valid email"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verified user and retrieve new token for filling personal data"
          },
          "400": {
            "description": "Invalid or expired token"
          }
        }
      }
    },
    "/api/auth/create": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Fill user's personal data",
        "description": "Fill verified user's entity with personal data",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "token",
                  "first_name",
                  "last_name",
                  "password"
                ],
                "properties": {
                  "token": {
                    "type": "string",
                    "format": "string",
                    "example": "aaaaabbbbbbccccc",
                    "description": "Token for managing process"
                  },
                  "first_name": {
                    "type": "string",
                    "format": "string",
                    "example": "John",
                    "description": "User's first name"
                  },
                  "last_name": {
                    "type": "string",
                    "format": "string",
                    "example": "Doe",
                    "description": "User's last name"
                  },
                  "password": {
                    "type": "string",
                    "format": "string",
                    "example": 12345678,
                    "description": "Password for gaining access to account"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User data is ready"
          },
          "400": {
            "description": "Invalid or expired token"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Log into account",
        "description": "Log into account by password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com",
                    "description": "Email for login"
                  },
                  "password": {
                    "type": "string",
                    "format": "string",
                    "example": 12345678,
                    "description": "Password for access"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verified user and retrieve new token for filling personal data"
          },
          "400": {
            "description": "Invalid or expired token"
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Refresh user access token",
        "description": "Refreshes the access token using a valid refresh token from cookies.",
        "responses": {
          "200": {
            "description": "Access token refreshed successfully."
          },
          "400": {
            "description": "Missing or invalid refresh token."
          },
          "403": {
            "description": "Refresh token is expired or unauthorized."
          }
        }
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Forgot password",
        "description": "Reset password by link.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com",
                    "description": "The email address to reset."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reset link will be sent to the given email."
          },
          "400": {
            "description": "Email is invalid or not registered."
          }
        }
      }
    },
    "/api/auth/change-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Change password",
        "description": "Change password based on token in reset link.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "token",
                  "password"
                ],
                "properties": {
                  "token": {
                    "type": "string",
                    "format": "string",
                    "example": "aaaaaaabbbbbbccccccc",
                    "description": "Token for update password."
                  },
                  "password": {
                    "type": "string",
                    "format": "string",
                    "example": 12345678,
                    "description": "Token for update password."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password updated successfully ."
          },
          "400": {
            "description": "Token is invalid."
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Log out from account",
        "description": "Log out from account with just one request.",
        "responses": {
          "200": {
            "description": "Logged out successfully."
          },
          "400": {
            "description": "Missing or invalid refresh token."
          },
          "403": {
            "description": "Refresh token is expired or unauthorized."
          }
        }
      }
    },
...
  },
...
}